requirements.txt
streamlit
openai
bcrypt
python-dotenv
sqlalchemy
requests
